app/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
jinja2
python-multipart
tenacity
requests
aiofiles
